# Conditional rate updates for MultiRate reactions