--- a/include/cantera/kinetics/ReactionData.h
+++ b/include/cantera/kinetics/ReactionData.h
@@ -37,10 +37,22 @@ struct ArrheniusData
     void update(double T, double P) { update(T); }
 
     //! Update data container based on *bulk* phase state
-    void update(const ThermoPhase& bulk, const Kinetics& kin);
+    //! @returns A pair where the first element indicates whether the `updateFromStruct`
+    //!      function for individual reactions needs to be called, and the second
+    //!      element indicates whether the `evalFromStruct` method needs to be called
+    //!      (assuming previously-calculated values were cached)
+    std::pair<bool, bool> update(const ThermoPhase& bulk, const Kinetics& kin);
 
     //! Update number of species and reactions; unused
     void resize(size_t n_species, size_t n_reactions) {}
+
+    //! Force shared data and reaction rates to be updated next time. This is called by
+    //! functions that change quantities affecting rate calculations that are normally
+    //! assumed to be constant, like the reaction rate parameters or the number of
+    //! reactions.
+    void invalidateCache() {
+        temperature = NAN;
+    }
 
     double temperature; //!< temperature
     double logT; //!< logarithm of temperature
@@ -55,8 +67,7 @@ struct ArrheniusData
  */
 struct BlowersMaselData
 {
-    BlowersMaselData()
-        : temperature(1.), logT(0.), recipT(1.), finalized(false) {}
+    BlowersMaselData();
 
     //! Update data container based on temperature *T*
     void update(double T);
@@ -67,7 +78,7 @@ struct BlowersMaselData
     }
 
     //! Update data container based on *bulk* phase state and *kin* kinetics
-    void update(const ThermoPhase& bulk, const Kinetics& kin);
+    std::pair<bool, bool> update(const ThermoPhase& bulk, const Kinetics& kin);
 
     //! Finalize setup
     void resize(size_t n_species, size_t n_reactions) {
@@ -76,9 +87,16 @@ struct BlowersMaselData
         finalized = true;
     }
 
+    void invalidateCache() {
+        temperature = NAN;
+    }
+
     double temperature; //!< temperature
     double logT; //!< logarithm of temperature
     double recipT; //!< inverse of temperature
+    double density; //!< used to determine if updates are needed
+    int state_mf_number; //!< integer that is incremented when composition changes
+
 
     bool finalized; //!< boolean indicating whether vectors are accessible
     vector_fp dH; //!< enthalpy change for each reaction
@@ -95,10 +113,10 @@ protected:
  */
 struct FalloffData : public ArrheniusData
 {
-    FalloffData() : finalized(false) {}
+    FalloffData() : finalized(false), molar_density(NAN), state_mf_number(-1) {}
 
     //! Update data container based on *bulk* phase state and *kin* kinetics
-    void update(const ThermoPhase& bulk, const Kinetics& kin);
+    std::pair<bool, bool> update(const ThermoPhase& bulk, const Kinetics& kin);
     using ArrheniusData::update;
 
     //! Finalize setup
@@ -107,8 +125,15 @@ struct FalloffData : public ArrheniusData
         finalized = true;
     }
 
+    void invalidateCache() {
+        ArrheniusData::invalidateCache();
+        molar_density = NAN;
+    }
+
     bool finalized; //!< boolean indicating whether vectors are accessible
     vector_fp conc_3b; //!< vector of effective third-body concentrations
+    double molar_density; //!< used to determine if updates are needed
+    int state_mf_number; //!< integer that is incremented when composition changes
 };
 
 
@@ -119,7 +144,7 @@ struct FalloffData : public ArrheniusData
  */
 struct PlogData
 {
-    PlogData() : temperature(1.), logT(0.), recipT(1.), logP(0.) {}
+    PlogData() : temperature(1.), logT(0.), recipT(1.), pressure(NAN), logP(0.) {}
 
     //! Update data container based on temperature *T* (raises exception)
     void update(double T);
@@ -129,18 +154,25 @@ struct PlogData
         temperature = T;
         logT = std::log(T);
         recipT = 1./T;
+        pressure = P;
         logP = std::log(P);
     }
 
     //! Update data container based on *bulk* phase state
-    void update(const ThermoPhase& bulk, const Kinetics& kin);
+    std::pair<bool, bool> update(const ThermoPhase& bulk, const Kinetics& kin);
 
     //! Update number of species and reactions; unused
     void resize(size_t n_species, size_t n_reactions) {}
 
+    void invalidateCache() {
+        temperature = NAN;
+        pressure = NAN;
+    }
+
     double temperature; //!< temperature
     double logT; //!< logarithm of temperature
     double recipT; //!< inverse of temperature
+    double pressure; //!< Pressure [Pa]
     double logP; //!< logarithm of pressure
 };
 
@@ -152,7 +184,7 @@ struct PlogData
  */
 struct ChebyshevData
 {
-    ChebyshevData() : temperature(1.), recipT(1.), log10P(0.) {}
+    ChebyshevData() : temperature(1.), recipT(1.), pressure(NAN), log10P(0.) {}
 
     //! Update data container based on temperature *T* (raises exception)
     void update(double T);
@@ -162,17 +194,24 @@ struct ChebyshevData
     {
         temperature = T;
         recipT = 1./T;
+        pressure = P;
         log10P = std::log10(P);
     }
 
     //! Update data container based on *bulk* phase state
-    void update(const ThermoPhase& bulk, const Kinetics& kin);
+    std::pair<bool, bool> update(const ThermoPhase& bulk, const Kinetics& kin);
 
     //! Update number of species and reactions; unused
     void resize(size_t n_species, size_t n_reactions) {}
 
+    void invalidateCache() {
+        temperature = NAN;
+        pressure = NAN;
+    }
+
     double temperature; //!< temperature
     double recipT; //!< inverse of temperature
+    double pressure; //!< Pressure [Pa]
     double log10P; //!< base 10 logarithm of pressure
 };
 
@@ -189,10 +228,14 @@ struct CustomFunc1Data
     void update(double T, double P) { update(T); }
 
     //! Update data container based on *bulk* phase state
-    void update(const ThermoPhase& bulk, const Kinetics& kin);
+    std::pair<bool, bool> update(const ThermoPhase& bulk, const Kinetics& kin);
 
     //! Update number of species and reactions; unused
     void resize(size_t n_species, size_t n_reactions) {}
+
+    void invalidateCache() {
+        temperature = NAN;
+    }
 
     double temperature; //!< temperature
 };