<!-- Thanks for contributing code! Please include a description of your change and check your pull request against the list below. For further questions, refer to the contributing guide (https://github.com/Cantera/cantera/blob/main/CONTRIBUTING.md). -->

**Changes proposed in this pull request**

<!-- Provide a clear and concise description of changes and/or features introduced in this pull request. -->

- Use the `RateData::update` method (executed once per reaction type) to determine whether or not both `updateFromStruct` and `evalFromStruct` need to be called for all reactions of that type.
- Remove other conditions that could cause evaluation of `BulkKinetics::m_bulk_rates` to be skipped incorrectly

**If applicable, provide an example illustrating new features this pull request is introducing**

The timing results from `custom_reactions.py` show an improvement compared to the "legacy" framework, with the new framework now outperforming the "legacy" one.

- Current `main` branch:
  - New framework (YAML): 257.54 ms (T_final=2736.60)
  - One Python reaction: 273.18 ms (T_final=2736.60) ... +6.07%
  - Old framework (XML): 253.31 ms (T_final=2736.60) ... -1.64%

- This PR:
  - New framework (YAML): 248.20 ms (T_final=2736.60)
  - One Python reaction: 268.20 ms (T_final=2736.60) ... +8.06%
  - Old framework (XML): 254.35 ms (T_final=2736.60) ... +2.48%

<!-- A minimal, complete, and reproducible example demonstrating features introduced by this pull request. See https://stackoverflow.com/help/minimal-reproducible-example for additional suggestions on how to create such an example. -->

**Checklist**

- [x] The pull request includes a clear description of this code change
- [x] Commit messages have short titles and reference relevant issues
- [x] Build passes (`scons build` & `scons test`) and unit tests address code coverage
- [x] Style & formatting of contributed code follows [contributing guidelines](https://github.com/Cantera/cantera/blob/main/CONTRIBUTING.md)
- [x] The pull request is ready for review